+++ src/libs/loaders/nodeSupervisor.ts
@@ -3,17 +3,21 @@ import {
     ChainConfig,
     NodeSignal,
     NodeSupervisor,
-    PipelineData,
     PipelineMeta,
     PipelineProcessor,
-    setMonitoringCallbacks,
-    setResolverCallbacks,
+    Ext,
     SupervisorPayloadDeployChain,
     SupervisorPayloadSetup,
+    PipelineData,
 } from 'dpcp-library';
 import { Logger } from '../loggers';
-import { InfrastructureService } from '../../utils/types/contractDataProcessing';
-import { nodeCallbackService } from '../../services/public/v1/node.public.service';
+import {
+    nodeCallbackService,
+    nodePreCallbackService,
+} from '../../services/public/v1/node.public.service';
+import { Service } from '../../utils/types/contractServiceChain';
+import { handle } from './handler';
+import axios from 'axios';
 
 export class SupervisorContainer {
     private static instance: SupervisorContainer;
@@ -90,7 +94,6 @@ export class SupervisorContainer {
                     this.nodeSupervisor.handleNotification(chainId, signal);
                     break;
                 }
-
                 default:
                     Logger.error({
                         message: 'Invalid communication type',
@@ -107,7 +110,15 @@ export class SupervisorContainer {
 
     public async setup(): Promise<void> {
         PipelineProcessor.setCallbackService(
-            async ({ targetId, data, meta }): Promise<PipelineData> => {
+            async ({
+                targetId,
+                data,
+                meta,
+                chainId,
+                nextTargetId,
+                previousTargetId,
+                nextNodeResolver,
+            }) => {
                 Logger.info({
                     message: `PipelineProcessor callback invoked - Connector: ${
                         this.uid
@@ -120,12 +131,53 @@ export class SupervisorContainer {
                     targetId,
                     data,
                     meta,
+                    chainId,
+                    nextTargetId,
+                    previousTargetId,
+                    nextNodeResolver,
                 });
             }
         );
 
-        await setResolverCallbacks({
+        PipelineProcessor.setPreCallbackService(
+            async ({
+                targetId,
+                data,
+                meta,
+                chainId,
+                nextTargetId,
+                previousTargetId,
+                nextNodeResolver,
+            }): Promise<PipelineData> => {
+                Logger.info({
+                    message: `${
+                        process.env.PORT
+                    } - PipelineProcessor pre callback invoked:
+                      - chainId: ${chainId}
+                      - nextTargetId: ${nextTargetId}
+                      - nextNodeResolver: ${nextNodeResolver}
+                      - Connector: ${this.uid}
+                      - Target: ${targetId}
+                      - MetaData: ${JSON.stringify(meta?.configuration)}
+                      - Data size: ${JSON.stringify(data)?.length} bytes
+                      - Received DATA: ${JSON.stringify(data ?? undefined)}
+          `,
+                });
+                return await nodePreCallbackService({
+                    targetId,
+                    data,
+                    meta,
+                    chainId,
+                    nextTargetId,
+                    previousTargetId,
+                    nextNodeResolver,
+                });
+            }
+        );
+
+        await Ext.Resolver.setResolverCallbacks({
             paths: {
+                pre: '/node/communicate/pre',
                 setup: '/node/communicate/setup',
                 run: '/node/communicate/run',
             },
@@ -148,7 +200,7 @@ export class SupervisorContainer {
             },
         });
 
-        await setMonitoringCallbacks({
+        await Ext.Reporting.setMonitoringCallbacks({
             paths: {
                 notify: '/node/communicate/notify',
             },
@@ -166,25 +218,42 @@ export class SupervisorContainer {
         }
     }
 
-    public processingChainConfigConverter(
-        dataProcessing: InfrastructureService,
-        participantEndpoint: string,
-        dataExchange?: string,
-        signedConsent?: any,
-        encrypted?: any
-    ): ChainConfig {
+    public async processingChainConfigConverter(props: {
+        serviceChain: Service;
+        participantEndpoint: string;
+        participantName: string;
+        participantCatalogId: string;
+        dataExchange?: string;
+        signedConsent?: any;
+        encrypted?: any;
+        isLast: boolean;
+    }): Promise<ChainConfig> {
+        const {
+            serviceChain,
+            participantEndpoint,
+            participantName,
+            participantCatalogId,
+            dataExchange,
+            signedConsent,
+            encrypted,
+            isLast,
+        } = props;
         const chainConfig: ChainConfig = [];
+
         chainConfig.push({
             chainId: '',
             services: [
                 {
-                    targetId: dataProcessing.serviceOffering,
+                    targetId: serviceChain.service,
                     meta: {
                         resolver: participantEndpoint,
                         configuration: {
-                            params: { ...dataProcessing.params },
+                            params: { ...serviceChain.params },
                             infrastructureConfiguration:
-                                dataProcessing.configuration,
+                                serviceChain.configuration,
+                            participantName,
+                            participantCatalogId,
+                            participantEndpoint,
                             dataExchange,
                             signedConsent,
                             encrypted,
@@ -193,7 +262,81 @@ export class SupervisorContainer {
                 },
             ],
             location: 'remote',
+            ...(isLast ? {} : { signalQueue: ['node_suspend'] }),
+            pre: await this.processPreChainConverter(
+                serviceChain.pre,
+                dataExchange,
+                signedConsent,
+                encrypted
+            ),
         });
         return chainConfig;
+    }
+
+    public async processPreChainConverter(
+        pre: any[],
+        dataExchange?: string,
+        signedConsent?: any,
+        encrypted?: any
+    ): Promise<
+        {
+            chainId: string;
+            location: string;
+            services: {
+                targetId: any;
+                meta: {
+                    resolver: string;
+                    configuration: {
+                        signedConsent: any;
+                        encrypted: any;
+                        infrastructureConfiguration: any;
+                        dataExchange: string;
+                        participantEndpoint: string;
+                        participantCatalogId: any;
+                        params: any;
+                        participantName: any;
+                    };
+                };
+            }[];
+        }[][]
+    > {
+        const finalArray = [];
+
+        for (const chain of pre) {
+            const subArray = [];
+            for (const service of chain) {
+                const [participantResponse] = await handle(
+                    axios.get(service.participant)
+                );
+                const participantEndpoint =
+                    participantResponse.dataspaceEndpoint;
+                subArray.push({
+                    chainId: '',
+                    services: [
+                        {
+                            targetId: service.service,
+                            meta: {
+                                resolver: participantEndpoint,
+                                configuration: {
+                                    params: { ...service.params },
+                                    infrastructureConfiguration:
+                                        service.configuration,
+                                    participantName: participantResponse.name,
+                                    participantCatalogId:
+                                        participantResponse._id,
+                                    participantEndpoint,
+                                    dataExchange,
+                                    signedConsent,
+                                    encrypted,
+                                },
+                            },
+                        },
+                    ],
+                    location: 'remote',
+                });
+            }
+            finalArray.push(subArray);
+        }
+        return finalArray;
     }
 }