+++ src/libs/loaders/representationFetcher.ts
@@ -5,6 +5,7 @@ import { paramsMapper } from '../../utils/paramsMapper';
 import { handle } from './handler';
 import { User } from '../../utils/types/user';
 import { getCredentialByIdService } from '../../services/private/v1/credential.private.service';
+import { Headers } from '../../utils/types/headers';
 
 /**
  * POST data to given representation URL
@@ -24,6 +25,10 @@ export const postRepresentation = async (params: {
     credential: string;
     decryptedConsent?: any;
     dataExchange?: IDataExchange;
+    chainId?: string;
+    nextTargetId?: string;
+    previousTargetId?: string;
+    nextNodeResolver?: string;
 }) => {
     const {
         method,
@@ -32,6 +37,9 @@ export const postRepresentation = async (params: {
         credential,
         decryptedConsent,
         dataExchange,
+        chainId,
+        nextTargetId,
+        previousTargetId,
     } = params;
 
     let cred;
@@ -44,6 +52,9 @@ export const postRepresentation = async (params: {
     const headers = headerProcessing({
         decryptedConsent,
         dataExchange,
+        chainId,
+        nextTargetId,
+        previousTargetId,
     });
 
     switch (method) {
@@ -91,6 +102,10 @@ export const putRepresentation = async (params: {
     credential: string;
     decryptedConsent?: any;
     dataExchange?: IDataExchange;
+    chainId?: string;
+    nextTargetId?: string;
+    previousTargetId?: string;
+    nextNodeResolver?: string;
 }) => {
     const {
         method,
@@ -99,6 +114,10 @@ export const putRepresentation = async (params: {
         credential,
         decryptedConsent,
         dataExchange,
+        chainId,
+        nextTargetId,
+        previousTargetId,
+        nextNodeResolver,
     } = params;
 
     let cred;
@@ -111,6 +130,10 @@ export const putRepresentation = async (params: {
     const headers = headerProcessing({
         decryptedConsent,
         dataExchange,
+        chainId,
+        nextTargetId,
+        previousTargetId,
+        nextNodeResolver,
     });
 
     switch (method) {
@@ -160,6 +183,10 @@ export const getRepresentation = async (params: {
     decryptedConsent?: any;
     representationQueryParams?: string[];
     dataExchange?: IDataExchange;
+    chainId?: string;
+    nextTargetId?: string;
+    previousTargetId?: string;
+    nextNodeResolver?: string;
 }) => {
     const {
         resource,
@@ -169,6 +196,10 @@ export const getRepresentation = async (params: {
         decryptedConsent,
         representationQueryParams,
         dataExchange,
+        chainId,
+        nextTargetId,
+        previousTargetId,
+        nextNodeResolver,
     } = params;
 
     let cred;
@@ -181,6 +212,10 @@ export const getRepresentation = async (params: {
     const headers = headerProcessing({
         decryptedConsent,
         dataExchange,
+        chainId,
+        nextTargetId,
+        previousTargetId,
+        nextNodeResolver,
     });
 
     let url;
@@ -244,6 +279,10 @@ export const postOrPutRepresentation = async (params: {
     user?: string;
     decryptedConsent?: any;
     dataExchange?: IDataExchange;
+    chainId?: string;
+    nextTargetId?: string;
+    previousTargetId?: string;
+    nextNodeResolver?: string;
 }) => {
     const {
         representationUrl,
@@ -253,6 +292,10 @@ export const postOrPutRepresentation = async (params: {
         verb,
         decryptedConsent,
         dataExchange,
+        chainId,
+        nextTargetId,
+        previousTargetId,
+        nextNodeResolver,
     } = params;
     // if contains params in URL is PUT Method
     if (representationUrl.match(Regexes.userIdParams)) {
@@ -271,6 +314,10 @@ export const postOrPutRepresentation = async (params: {
                 credential,
                 decryptedConsent,
                 dataExchange,
+                chainId,
+                nextTargetId,
+                previousTargetId,
+                nextNodeResolver,
             })
         );
 
@@ -290,6 +337,10 @@ export const postOrPutRepresentation = async (params: {
                 credential,
                 decryptedConsent,
                 dataExchange,
+                chainId,
+                nextTargetId,
+                previousTargetId,
+                nextNodeResolver,
             })
         );
 
@@ -306,6 +357,10 @@ export const postOrPutRepresentation = async (params: {
                         data,
                         credential,
                         dataExchange,
+                        chainId,
+                        nextTargetId,
+                        previousTargetId,
+                        nextNodeResolver,
                     })
                 );
 
@@ -319,6 +374,10 @@ export const postOrPutRepresentation = async (params: {
                         data,
                         credential,
                         dataExchange,
+                        chainId,
+                        nextTargetId,
+                        previousTargetId,
+                        nextNodeResolver,
                     })
                 );
 
@@ -332,6 +391,10 @@ export const postOrPutRepresentation = async (params: {
                         data,
                         credential,
                         dataExchange,
+                        chainId,
+                        nextTargetId,
+                        previousTargetId,
+                        nextNodeResolver,
                     })
                 );
 
@@ -352,12 +415,35 @@ export const postOrPutRepresentation = async (params: {
 const headerProcessing = (params: {
     decryptedConsent?: any;
     dataExchange?: IDataExchange;
+    chainId?: string;
+    nextTargetId?: string;
+    previousTargetId?: string;
+    nextNodeResolver?: string;
 }): object => {
-    const { decryptedConsent, dataExchange } = params;
+    const {
+        decryptedConsent,
+        dataExchange,
+        chainId,
+        nextTargetId,
+        previousTargetId,
+        nextNodeResolver,
+    } = params;
 
-    let headers = {};
+    let headers: Headers = {
+        ...(chainId ? { 'x-ptx-service-chain-id': chainId } : {}),
+        ...(nextTargetId
+            ? { 'x-ptx-service-chain-next-target': nextTargetId }
+            : {}),
+        ...(previousTargetId
+            ? { 'x-ptx-service-chain-previous-target': previousTargetId }
+            : {}),
+        ...(nextNodeResolver
+            ? { 'x-ptx-service-chain-next-node': nextNodeResolver }
+            : {}),
+    };
     if (decryptedConsent) {
         headers = {
+            ...headers,
             consentId: decryptedConsent?._id,
             'consent-id': decryptedConsent?._id,
             'x-ptx-consent-id': decryptedConsent?._id,
@@ -372,7 +458,7 @@ const headerProcessing = (params: {
             'x-ptx-incomingDataspaceConnectorURI': dataExchange.consumerEndpoint
                 ? dataExchange.consumerEndpoint
                 : dataExchange.providerEndpoint,
-            'x-ptx-dataExchangeId': dataExchange?._id,
+            'x-ptx-dataExchangeId': dataExchange?._id.toString(),
             'x-ptx-contractId': dataExchange.contract.split('/').pop(),
             'x-ptx-contractURL': dataExchange.contract,
         };