+++ src/services/public/v1/data.public.service.ts
@@ -198,15 +198,14 @@ export const exportDataService = async ({
                     );
 
                     //When the data is retrieved, check wich flow to trigger based infrastructure options
-                    // the options and the dataProcessings will be represented in the consent
+                    // the options and the serviceChains will be represented in the consent
                     if (
-                        dataExchange.dataProcessing &&
-                        dataExchange.dataProcessing.infrastructureServices
-                            .length > 0
+                        dataExchange.serviceChain &&
+                        dataExchange.serviceChain.services.length > 0
                     ) {
                         //Trigger the infrastructure flow
                         await triggerInfrastructureFlowService(
-                            dataExchange.dataProcessing,
+                            dataExchange.serviceChain,
                             dataExchange,
                             data,
                             signedConsent,