+++ src/services/public/v1/provider.public.service.ts
@@ -113,14 +113,13 @@ export const ProviderExportService = async (
 
                     //When the data is retrieved, check wich flow to trigger based infrastructure options
                     if (
-                        dataExchange.dataProcessing &&
-                        dataExchange.dataProcessing.infrastructureServices
-                            .length > 0
+                        dataExchange.serviceChain &&
+                        dataExchange.serviceChain.services.length > 0
                     ) {
                         //Trigger the infrastructure flow
 
                         await triggerInfrastructureFlowService(
-                            dataExchange.dataProcessing,
+                            dataExchange.serviceChain,
                             dataExchange,
                             data
                         );
@@ -135,6 +134,12 @@ export const ProviderExportService = async (
                             endpointData,
                         });
                     }
+                    Logger.info({
+                        message: `Successfully retrieve data from ${resourceSD} with size of ${
+                            JSON.stringify(data).length
+                        }Bytes`,
+                        location: 'ProviderExportService',
+                    });
                 }
             }