+++ src/utils/types/contractServiceChain.ts
@@ -1,0 +1,15 @@
+export type Service = {
+    participant: string;
+    service: string;
+    configuration: string;
+    params: {
+        [key: string]: string;
+    };
+    completed?: boolean;
+    pre?: any[];
+};
+
+export type ContractServiceChain = {
+    catalogId: string;
+    services: Service[];
+};